initial addition of ammo resources
- also refactor of model resource functions to separate file
