+++ game/resources/weapons/ballistic/cl_lbx_ac_20.yaml
@@ -3,6 +3,7 @@
 name: LB 20-X AC
 short: LBxAC20
 tech: clan
+audio: autocannon.ogg
 tonnage: 12
 damage: 20
 heat: 6
@@ -12,7 +13,7 @@ distance: 360
 velocity: 350
 # cooldown in seconds
 cooldown: 1.75
-audio: autocannon.ogg
+ammoPerTon: 5
 projectileCount: 20
 projectileSpread: 0.025
 projectile:
