+++ game/resources/weapons/missile/cl_lrm_10.yaml
@@ -3,6 +3,7 @@
 name: Long Range Missile 10
 short: LRM10
 tech: clan
+audio: missile-0.ogg
 tonnage: 2.5
 damage: 10
 heat: 4
@@ -12,7 +13,7 @@ distance: 630
 velocity: 150
 # cooldown in seconds
 cooldown: 4
-audio: missile-0.ogg
+ammoPerTon: 120
 lockOn:
   turnRate: 7.5
   groupRadius: 0.15
